move isValidDate, convertTimestampToDate and renderWebsites to GameUtils
